Fix JStudio not compiling on retail, add fake header
