+++ include/JSystem/JStudio/JStudio_JStage/control.h
@@ -10,6 +10,15 @@
 #include "JSystem/JStudio/JStudio/jstudio-object.h"
 #include "JSystem/JStudio/JStudio/jstudio-math.h"
 
+#ifdef DEBUG
+namespace JStudio_JStage {
+template<class TAdaptor, class TStageObject>
+struct TVariableValueOutput_object_;
+};  // namespace JStudio_JStage
+#else
+#include "JSystem/JStudio/JStudio_JStage/tvariable_value_output_object.h"
+#endif
+
 namespace JStudio_JStage {
 typedef JStudio::TObject* (*ObjCreateFuncT)(const JStudio::stb::data::TParse_TBlock_object&, JStage::TObject*, const JStage::TSystem*);
 
@@ -40,10 +49,6 @@ struct TAdaptor_object_ {
     /* 0x0 */ JStage::TSystem const* pJSGSystem_;
     /* 0x4 */ JStage::TObject* pJSGObject_;
 };
-
-
-template<class TAdaptor, class TStageObject>
-struct TVariableValueOutput_object_;
 
 struct TAdaptor_actor : public JStudio::TAdaptor_actor, public JStudio_JStage::TAdaptor_object_ {
     typedef JStudio::TObject_actor ObjectType;
@@ -282,35 +287,6 @@ struct TAdaptor_light : public JStudio::TAdaptor_light, public TAdaptor_object_ 
     static TVVOutput_direction_ saoVVOutput_direction_[6];
 };
 
-template<class TAdaptor, class TStageObject>
-struct TVariableValueOutput_object_ : public JStudio::TVariableValue::TOutput {
-    typedef f32 (TStageObject::*GetFunc)() const;
-    typedef void (TStageObject::*SetFunc)(f32);
-    TVariableValueOutput_object_() : field_0x4(-1), field_0x8(NULL), field_0x14(NULL) {}
-    TVariableValueOutput_object_(typename TAdaptor::TEVariableValue param_1, 
-    SetFunc param_2, GetFunc param_3) : field_0x4(param_1), field_0x8(param_2), field_0x14(param_3) {
-
-    }
-
-    virtual void operator()(f32 param_1, JStudio::TAdaptor* param_2) const {
-        (((TAdaptor*)param_2)->get_pJSG_()->*field_0x8)(param_1);
-    }
-    virtual ~TVariableValueOutput_object_() {}
-
-    bool isEnd_() const { return field_0x4 == -1; }
-    void adaptor_setOutput_(TAdaptor* adaptor) {
-        adaptor->adaptor_referVariableValue(field_0x4)->setOutput(this);
-    }
-    void setVariableValue_(const TStageObject* pObj, TAdaptor* pAdaptor) const {
-        f32 val = (pObj->*field_0x14)();
-        pAdaptor->adaptor_setVariableValue_immediate(field_0x4, val);
-    }
-
-    int field_0x4;
-    SetFunc field_0x8;
-    GetFunc field_0x14;
-};
-
 /* 8028A1F8 */ bool
     transform_toGlobalFromLocal(f32 (*)[4],
                                 JStudio::TControl::TTransform_translation_rotation_scaling const&,
@@ -338,5 +314,9 @@ inline bool transform_toGlobalFromLocal(JStudio::TControl::TTransform_position* 
                                                 JStudio::TControl::TTransform_position const&,
                                                 JStage::TObject const*, u32);
 };  // namespace JStudio_JStage
+
+#ifdef DEBUG
+#include "JSystem/JStudio/JStudio_JStage/tvariable_value_output_object.h"
+#endif
 
 #endif /* JSTUDIO_JSTAGE_CONTROL_H */
